feat: Add app to generate a Lighthouse report for an entire site
The metrics app runs a Lighthouse report for every page loaded from
the site's sitemap (recursing through site indexes). The data for
each page is collected and published as a report in PDF format.

For each category: peformance, accessibility, best practices and seo
(pwa is excluded for now), tables showing the number of pages in each
rating: poor, needs improvement and good are displayed, along with
distribution graphs of the number of pages in percentile group (5%)
and tables showing the 20 urls with the lowest scores.

For the performance category, the same results are shown for the three
main audits: largest-contentful-paint, total-blocking-time and
cumulative-layout-shift which are the main metrics used to determine the
performance score for a page, and the ones which are used in detemrining
the site's ranking.